requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djlint==1.0.0
idna==3.3
importlib-metadata==4.11.3
pathspec==0.9.0
PyJWT==2.3.0
pysbd==0.3.4
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomlkit==0.10.2
tqdm==4.64.0
tzdata==2022.1
urllib3==1.26.9
Wikipedia-API==0.5.4
zipp==3.8.0
